+++ app/database/queries/weather_queries.py
@@ -20,9 +20,9 @@ UPSERT_WEATHER_HOURLY = """
         breach_metric, breach_observed_value, breach_threshold_value,
         threshold_id, cycle_id
     ) VALUES (
-        $1::uuid, $2, $3, $4, $5, $6::numeric, $7::numeric,
+        $1::uuid, $2, $3, $4, $5, $6::float8, $7::float8,
         ST_SetSRID(ST_MakePoint($7::float8, $6::float8), 4326),
-        $8, $9::numeric, $10, $11, $12, $13, $14, $15, $16, 
+        $8, $9, $10, $11, $12, $13, $14, $15, $16, 
         $17, $18, $19, $20, $21, $22, $23, $24,
         $25, $26, $27, $28, $29, $30, $31, $32, 
         $33, $34, $35, $36, $37, $38, $39, $40, $41, $42,
@@ -85,7 +85,7 @@ UPSERT_WEATHER_DAILY = """
         flag_heavy_rain_day, flag_storm_day, flag_cold_wave_day,
         worst_breach_severity, cycle_id
     ) VALUES (
-        $1::uuid, $2, $3, $4, $5, $6::numeric, $7::numeric,
+        $1::uuid, $2, $3, $4, $5, $6::float8, $7::float8,
         ST_SetSRID(ST_MakePoint($7::float8, $6::float8), 4326),
         $8, $9, $10, $11, $12, $13, $14, $15, 
         $16, $17, $18, $19, $20, $21, $22, $23, $24, $25, $26, $27::uuid
